kpromo: move version to one level up and create rules for building each app
