+++ Dockerfile
@@ -42,7 +42,7 @@ RUN make kpromo
 FROM gcr.io/google.com/cloudsdktool/cloud-sdk:slim AS base
 
 WORKDIR /
-COPY --from=builder /go/src/app/kpromo .
+COPY --from=builder /go/src/app/bin/kpromo .
 
 ENTRYPOINT ["/kpromo"]
