+++ cmd/kpromo/cmd/root.go
@@ -25,6 +25,7 @@ import (
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/cip"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/manifest"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/run"
+	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/version"
 	"sigs.k8s.io/release-utils/log"
 )
 
@@ -74,6 +75,7 @@ func init() {
 	rootCmd.AddCommand(run.RunCmd)
 	rootCmd.AddCommand(cip.CipCmd)
 	rootCmd.AddCommand(manifest.ManifestCmd)
+	rootCmd.AddCommand(version.VersionCmd)
 }
 
 func initLogging(*cobra.Command, []string) error {
